requirements.txt
torch>=1.9.0
torchvision>=0.10.0
gradio>=3.40.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.3.0
pathlib
tqdm
matplotlib
transformers>=4.20.0
timm>=0.6.0
pycocotools
opencv-python-headless
av
scipy
scikit-image
iopath